kgr: expose global 'in_progress' state through sysfs
In addition to having a per-process flag that shows which processess have
already been "migrated", it's useful to have a global-wide flag that will
show whether the patching operation is currently undergoing without having
to traverse all /proc entries.

[v2]
Switch from procfs to sysfs as we want to have more entries there
later. So we create a one-entry attribute group and register it under
/sys/kernel/kgraft.

js: handle error
js: move to kgraft_files.c, there will be more of them, do not mess
    kgraft.c, i.e. the core

Reported-by: Libor Pechacek <lpechacek@suse.cz>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>