kgr: support revert of patches
Add support to remove a module with patch. This effectively reverts
the patch and allows the module removal. We do this by adding a
'revert' file to patch's sysfs directory. A write to that file starts
the process. After it is finished, the module refcount is lowered and
the module can be removed.

We do not do this in the module rmmod path for two reasons:
* exit fucntion of a module cannot fail. This means, kgraft revert
  could not fail. But bad things happen. Removal of a module, when the
  revert fails would be disaster. Waiting infinitely in such case is
  not nice at all.
* Having processes waiting in the D state on rmmod path for kGraft to
  finish is not nice for the very same reason, even though it is only
  tens of seconds.

Signed-off-by: Jiri Slaby <jslaby@suse.cz>