kgr: avoid potential races in kgr_finalize()
Skipped patches need to be applied when the related modules are being loaded.
The patch also need to be disabled when the patched module is being removed.
Both these operations need to manipulate also the patch in progress that has not
been finalized yet.

This commit makes sure that status flags are not manipulated without
kgr_in_progress_lock.

Also it makes sure that kgr_patch is NULL when the patch is finalized.
It prevents from processing the patch as non-finalized.

Finally, it slightly reorganizes the calls to safe one check and set
both "NULL" and "false" together. It looked better readable to me.

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>