kgr: try to apply skipped patches when a module is loaded
kGraft patches might modify both core kernel functions and modules. To
make the life easier, new functions for modules have the
"abort_if_missing" flag disabled. The result is that such patch could
be loaded even before the affected module.

This commit adds kgr_module_init() function that is called when the
module is being loaded but before the code is used. It allows to apply
the fast path from the pending patches.

The support for handling the patch in progress will be added later. It
will need more tweaking.

Also the support for more patches against the same function will be
added later.

js: fix !MODULES configs

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>