kgr: handle patched modules that are being removed
kGraft allows to patch kernel modules. Such functions typically have disabled
the flag "abort_on_error". It allows to load the patch even before the affected
module.

This commit solves the situation when a patched module is being removed.
The patch might stay but the obsolete ftrace filters must be removed and
the status must be updated.

Everything has to be ready to patch the module when it is eventually loaded
back again.

The error handling is not ideal and will be improved in the future.

Also the support for handling the patch in progress will be added later. It will
need more tweaking.

Finally, the support for more patches against the same function will be added
later.

js: make kgr_module_exit_nb static (reported by Fengguang Wu
    <fengguang.wu@intel.com>)
js: fix !MODULES builds
pm: do not mess error codes when patching modules (remove "err |=")

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>