kgr: handle module load and removal for non-finished patches
We need to be more careful when a module is modified by a patch in progress.
kgr_finalize() still needs to be called and it expects the slow stub.

This commit solves the module load by introducing another state
KGR_PATCH_APPLIED_NON_FINALIZED. It means that the fast stub is already
in place but the finalize has not been called for the entire patch yet.

Note that the patch is progress is not be applied when it is being removed.
There is nothing to add in this situation :-)

The module removal is easy. The target state is "KGR_PATCH_SKIPPED" and
it is ignored by kgr_finalize().

pm: do not mess error codes when patching modules (remove "err |=")

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>