kgr: remove patch from global list when being removed
There is a problem when skipped patches are applied to modules.
We must ignore patches that are being removed. But this does not work now
because the patch is in the global list until it is fully removed.

This commit solves the problem by removing the patch from the global list
immediately when the revert begins.

It means that the global list contains only finalized patches.

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>