kgr: confusion with fatal state
The patch state is currently shown the following way:

    $> cat /sys/kernel/kgraft/sample_patcher/state
    Function              State  Fatal
    SyS_iopl                  1      1
    capable                   1      1
    unknown_function          5      0
    ftr_test_filter_read      1      0

I am always scared when I see that some functions have the fatal set.
It has somehow negative meaning about that something went wrong very badly.
This patch changes the output to:

    $> cat /sys/kernel/kgraft/sample_patcher/state
    Function               Weak  State
    SyS_iopl                  0      1
    capable                   0      1
    unknown_function          1      5
    ftr_test_filter_read      1      1

The meaning is that the symbol is weak. We even might change the
"abort_if_missing" variable to "weak" and revert the condition
everywhere.

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Acked-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>