kgr: make it possible to avoid lazy switching
In cases where the patch is known to be completely independent on other
context (such as, but not limited to, added check at the beginning of a
syscall), it's convenient to be able to completely avoid going through the
hoops of lazy migration to the new universe.

Allow patches to set an 'immediate' flag and not (re-)start the lazy
migration when it's set.

jk: simplify 'immediate' logic (remove preparation to parallelism)
js: call kgr_finalize properly also with immediate flag set

Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>