kgr: fix patch stacking also when loading module
If a module symbol is patched multiple times, only the last ftrace_ops
must be registered. Otherwise, the first registered stub is used and
the later patches are ignored.

This solution is suboptimal because even obsolete patches are registered and
immediately unregistered. On the other hand, the operation is not time critical
and the current solution allows to reuse the existing code.

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>