kgr: fix ugly race
The code contains an ugly race. If you are familiar with this notion:
(kgr_mark_processes . for_each_patch(kgr_patch_code)) || kgr_mark_safe

I.e. informally, when we set the in_progress bit for a process, it can
immediately clear the flag, but still calls the old function, because
we have not installed the slow stub via kgr_patch_code yet. When we
suddenly install it, the process can be in the middle of something
(and not safe to switch at all).

To fix that, we have to introduce a three-state logic:
* old-world-immutable
* old-world
* new-world

This is accomplished by introducing an atomic bit which we set prior
marking processes (old-world-immutable state) and clear after
installing the slow stub (old-world state). The slow stub always
re-sets the in_progress flag to 1 when in the old-world-immutable
universe. So only the exit from the syscall allows the switch to the
new-world.

jk: describe kgr_immutable
js: {
	Fix the race properly.

	First, set_bit before mark_processes, because this avoid race
	between them and slow stub.

	Second, merge kgr_mark_processes into kgr_handle_processes.
	Hence do the stuff after the stub is in place. This avoids
	another race.

	Finally, add memory barriers to force the consistent state.
	This is needed because, set_bit et al. are not implicit
	barriers (on non-x86 HW at least).

	Reported-by: Li Bin <huawei.libin@huawei.com>
}

Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>
Cc: Vojtech Pavlik <vojtech@suse.cz>
Cc: Petr Mladek <pmladek@suse.cz>