kgr: clean up reference handling when a patch replaces all others
I was always a bit confused by "p->refs--" in kgr_replace_all().
It decreases reference for each patch in kgr_patches list. In fact,
it does the same thing as kgr_refs_dec() and is just an optimization.

The reference numbers define dependency between patches. They help to make
sure that the patches are reverted in the right order.

This patch causes that if a patch is going to replace all older patches,
it will not increment the references in the first place.

Note that this patch should not cause any change in the behavior. All changes
to the older patches are blocked until the last patch is finalized. It is guarded
by the kgr_in_progress variable.

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>