kgr: correctly remove replaced patch when using fast path
kgr_replace_all() allows to remove a patch immediately when it is not longer
used. It currently just puts the related module. But it should also remove
the patch from the kgr_patches list and decrease reference of the dependent
patches.

This patch adds kgr_refs_dec_limited() function. It decreases the reference
for all patches that were added before the given one.

Also it adds kgr_remove_patch_fast() that does all the remaining steps
to remove the patch from any use.

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Reviewed-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>