kgr: lower the workqueue scheduling timeout
After a 'kgr poke', one has to wait 30s before the next patch can be
applied. The tools do not take that into account. As a workaround, I'm
going to add a busy wait to the %pre script, but the timeout should be
reduced to something more reasonable.

This use case uis quite realistic:
1) Admin applies patch1, forgets to check the progress
2) Admin applies patch2, this fails, because some tasks are still sleeping in
   the old universe
3) Admin does 'kgr poke' to wake them up
4) Admin applies patch2, this takes 30 more seconds to actually start doing
   anything (currently, it fails)

(quote from bsc#905087)

Fix this by lowering the timeout to 2 seconds. The overhead this adds
should be negligible -- we are scheduling the workqueue only if the
patching is still in progress, and the only thing that workqueue handler
does is that it walks the list of existing processess.

While at it, make the "still patching" message appear only once per
patching (it will be followed by "succeeded" message later, so the fact
that the patching hasn't been finalized yet will be obvious from dmesg
anyway) to avoid excessive spamming of the kernel log.

Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>