kgr: allow to search various types of struct kgr_patch_fun
kgr_get_last_pf() returns the last variant of struct kgr_patch_fun
that has set KGR_PATCH_APPLIED state. We will need to search and
check also another variants of the function on the stack in
the upcoming patches.

This patch adds the needed support by renaming kgr_get_last_pf() to
kgr_get_patch_fun() and extending the functionality. It looks
for struct patch_fun by the information in the given struct
kgr_patch_fun. In most cases only the function name is used.
The right search mode is defined by a new enum kgr_find_type.
The currently supported types are:

  + KGR_PREVIOUS: look for the previous patch_fun in respect to
    stacking; it searches both finalized patches (kgr_patches) and
    also the patch in progress (kgr_patch); the patch in progress
    is considered to be on top of the stack

  + KGR_LAST_FINALIZED: look for the last finalized patch_fun;
    it searches only the patches in kgr_patches list

  + KGR_LAST_EXISTING: look for the last existing patch_fun;
    it searches both kgr_patch and kgr_patches

  + KGR_IN_PROGRESS: look if the function in the given patch_fun
    is modified also by the patch in progress; it searches
    _only_ kgr_patch

Note that the patch modifies the meaning of the function but not the
behavior. Regarding the meaning:

  + kgr_get_last_pf() looked for the last applied patch_fun and checked
    the state KGR_PATCH_APPLIED.

  + kgr_get_patch_fun() ignores states and is interested only into the patch
    order. The search for KGR_PREVIOUS patch_fun stops when it meets the given
    patch_fun, the check is done by pointer to the struct kgr_patch_fun.

Regarding the behavior, the function is currently used only in
four situations:

  + initializing fops for new patch; it is used only for existing
    symbols that are completely new or already applied; it means
    the check for KGR_PATCH_APPLIED always passes; it will be
    the same for the pointer check because the struct is new

  + looking for old_ops when modifying patches. Also here it is
    called only when KGR_PATCH_APPLIED always passes; it will
    be the same with pointer because the pointer won't be
    accessible from kgr_patches

  + initializing fops in kgr_handle_patch_for_loaded_module();
    we strictly need the previous patch here; it works by chance
    because we set the KGR_PATCH_APPLIED state in the patches
    in the right order; this is the tricky usage that could
    break with further changes

  + looking for old fops in kgr_handle_patch_for_loaded_module();
    same situation as in the previous case

Finally, note that only the KGR_PREVIOUS type is used now.

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Acked-by: Jiri Kosina <jkosina@suse.cz>
Reviewed-by: Miroslav Benes <mbenes@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>