kgr: be consistent when applying patches on loaded modules
We must not use the fast stub when applying patch in progress
on the loaded module. The patched functions from the module
might call another patched functions from the kernel core.
Some threads might still use the old code for the kernel core
and thus the calls would be inconsistent.

We need to use the slow stub and let it decide what universe
is the right one.

This patch adds new function kgr_is_patch_fun(). It allows to check if the
given struct kgr_patch_fun is of the expected type. The types are the same
that are used to search by kgr_get_patch_fun().

The new function is then used to decide about the patch state and about
the registered ftrace ops when patching the loaded module.

Note that we always have to initialize the addresses and set the right state.
But we need to register the ftrace ops only for the last patch.

Also note that we need to handle the patch in progress even when it is being
reverted. The module must use the obsolete functions in threads that still
use the obsolete code for the rest of the kernel. By other words, all functions
in the module and core kernel must come from the same universe.

The change helps to get rid of the crazy KGR_PATCH_APPLIED_NON_FINALIZED
state. Also it is more effective because we register the ftrace ops only once.
Note that we have searched the previous patches even before when looking for
the unreg_ops.

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Reviewed-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>