kgr: handle IRQ context using global variable
kgr_revert_replaced_funs() calls kgr_patch_code() for finalized patches.
It sets the slow stub and allows a correct revert of obsolete patches.

The problem is that kgr_stub_slow() accesses this_cpu_ptr(p->patch->irq_use_new)
and this variable is available only for the patch is progress.

In fact, the state of the IRQ context is not specific to a particular patch.
It should be handled by a global variable.

This patch takes the minimalist approach and just replaces the patch-specific
variable with a global one. It is still allocated when a patch is manipulated
and freed when the patch is finalized. It has the advantage that we do not
take the per-CPU memory when not needed. Also we do not need to reset it
when patching starts.

I also thought about renaming the variable to kgr_irq_in_progress. In fact,
it has the same meaning as the TIF_KGR_IN_PROGRESS flag. But it would revert
the logic and we would need to initialize the value to true. Therefore
I kept it as is for now.

Signed-off-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>