kgr: correctly handle the first patching stage
The patch in progress (kgr_patch) is always applied in two stages. First
the slow stubs are registered via kgr_patch_code call.  Second the patch
is finalized and slow stubs are replaced by the fast ones.

If we fail during the first stage we need to go through already
processed patch_funs and replace the registered slow stubs with the
previously registered fast stubs (i.e. inverse action to
kgr_patch_code). Similarly for the reversion.

The situation differs for the replace_all patch. It can fail during the
reversion of stacked functions or during the very application of the
patch.  In both cases we need to process all the patches in the stack
and act accordingly.

We achieve all this by simply calling kgr_patching_failed function.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>