kgr: allow to disable patching
During the patching we could get to error situations which cannot be
solved by returning to the consistent state. The best policy available
is to disable further patching completely. This patch adds such
capability to existing kGraft code.

New global kgr_enabled boolean is defined with its sysfs interface.
Checks for kgr_enabled are added to all necessary places.

Delayed loading in disabled state is allowed only for modules which do
not contain any function to be patched. For the going module we cannot
prevent unloading so we warn the user.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>