kgr: handle the delayed patching of the modules
Also the delayed patching of the modules can fail. When the insertion of
the module fails we refuse to load it. However the user can force the
insertion in spite of the errors by setting force_load_module in sysfs
infrastructure. When true (default is false) the errors are omitted and
the module is loaded.

If an error occurs for going module we cannot do anything else than to
disable further patching. Otherwise consistent state could not be
guaranteed since we cannot prohibit the module to go away.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>