Revert "kgr: x86: optimize handling of CPU-bound tasks"
This reverts commit fa6046cd5bed46c170fef212ad60135f8366dd8d.

Looking at the stack of the task that is running is not guaranteed to
yield consistent result. The current implementation is very likely "fine",
in a sense that nr_entries will always be >2 even in cases when the actual
stack contents returned would be bogus (due to the stack undergoing active
changes at the time of dumping), but we'd better be on a safe side and
don't rely on undocumented and non-guaranteed behavior of stack dumping.

It's an optimization that can be easily sacrificed for now anyway.

Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>