kgr: take switching of the fops out of kgr_patch_code to new function
There is switching of the ftrace_ops in kgr_patch_code. New_ops are
enabled and the old ones disabled afterwards. Lets put it into the new
function kgr_switch_fops.

Also we change the error handling a bit. Currently we do not return
error if unregistration of the ftrace stub fails in kgr_patch_code
function. The process proceeds and the user is warned. However the
system gets to very suspicious state, because the ftrace stub could
still be there or not at all. There is nothing we can do about it, so we
better BUG here.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Petr Mladek <pmladek@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>