x86/kgr: move kgr infrastructure from asm to C
There are several places in entry_64.S where we need to clear
TIF_KGR_IN_PROGRESS flag in order to migrate the task during patching.
We can modify TIF work masks (that is _TIF_WORK_SYSCALL_ENTRY,
_TIF_WORK_MASK, _TIF_ALLWORK_MASK and _TIF_DO_NOTIFY_MASK) to achieve
the same in C.

We also gain one more place of migration with this as a bonus. The tasks
can be safely migrated on return from an interrupt/exception back to the
userspace. It is going to be essential for the fake signal.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>