kgr: remove immediate feature
kGraft supports immediate feature which, if set for a patch, bypass a
consistency model and the patch is applied immediately as is. It can be
used for number of fixes which does not require special approach. The
patching is finished in an instant and there is no need to handle
blocking tasks.

It has one major drawback though. If an immediate patch is reverted or
an immediate replace-all patch is applied it is done right away and a
reference to a module with the patch is put. Thus the module can be
removed. However there is no guarantee that there is no task running or
sleeping in the removed code. This could easily cause serious problems.
For this reason the immediate feature cannot be used anymore and we can
remove it.

The consistency model is now the only way how to apply a patch. It
guarantees that when the patching is finished there is no task in old
code which can then be safely removed. The problem of blocking tasks is
substantially alleviated with the introduction of the fake signal.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>