s390/kgr: reorganize kgr infrastructure in entry64.S
Currently we directly clear TIF_KGR_IN_PROGRESS in several places of
s390 entry64.S. Let's move the TIF to a different bit and take
advantage of _TIF_WORK_SVC and _TIF_WORK_INT masks which define the
work to be done during the returns from a syscall or an
interrupt/exception. Task migration is now done in a C function called
from assembly.

Since there are two ways how the syscall can be restarted on return
from a signal handling process, it is important to clear the flag
before do_signal() is called. Otherwise we could miss the migration if
we used SIGSTOP/SIGCONT procedure or fake signal to migrate patching
blocking tasks. If we place our hook to sysc_work label in entry
before TIF_SIGPENDING is evaluated we kill two birds with one stone.
The task is correctly migrated in all return paths from a syscall.

KGraft infrastructure corresponds to x86 implementation now. Our TIF
can be cleared on syscall entry/exit or on interrupt/exception return
to userspace.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>