kgr: fix migration of kthreads to the new universe
Currently when all the slow stubs are applied all tasks in the system
are marked as being in progress of migration to the new universe. If the
task is a kthread it is immediately woken up with wake_up_process()
function in order to being migrated.

Unfortunately there is a problem with this approach. If the kthread
calls a patched function between its awakening and its call to
schedule() it appears in kgraft slow stub. There the redirection to
the correct function should happen. If kgr_immutable flag is set
(meaning that slow stub application is not finished yet during current
patching process) the task is marked as non-migrated and its
TIF_KGR_IN_PROGRESS flag is set again. This prevents certain races which
could cause corruption. The issue is that the flag could be cleared
earlier by the kthread itself and it is set again in the slow stub.  It
will not be in the future anymore. Such kthread would block the patching
process.

Fix it by moving wake_up_process() for all the kthreads after the
kgr_immutable is cleared.

Note that there is no corruption involved. It is by all means possible
that an affected kthread could wake up and clear its flag by itself.
However in general kthreads can block the patching process infinitely.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>