kgr: wake up kthreads periodically
We call wake_up_process() for kthreads only once during the patching (or
after it to be precise). Unfortunately this is not sufficient. A kthread
is marked as non-migrated (its TIF_KGR_IN_PROGRESS is set) and
wake_up_process() is called on this kthread after the patching. However
if the kthread is running, wake_up_process() has no effect and the
kthread can go to sleep without a call to kgr_task_safe(). Thus the
kthread would block the patching process from finishing till it is woken
up again by some other process than kgraft. We can fix this by calling
wake_up_process() periodically in the same way we do it with sending a
fake signal.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>
Reviewed-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>