kgr: Remove the confusing search for fentry
First, the name of ftrace_function_to_fentry() function is confusing.
It returns the address that is modified by ftrace. It is the location
where the function gets redirected to the ftrace handler. The location
and its name is architecture specific. For example, ftrace might use
ftrace or mcount location on x86.

Second, the function was introduced at times when we tried to get kGraft
working with the ftrace handled called from the mcount location.
The function was redirected after the prologue. But the prologue
sometimes was incompatible between the original and the new function.
In this case, the redirection did not work and the system went down.

By other words kGraft works reliably only when the ftrace handler
is called at the beginning of the patched function[*] and the new
function is called from the beginning.

This patch removes the confusing function and all its calls.
The function address is used instead.

It removes the cache for the new function location. The value
is the same as the user-provided function pointer.

The term "function address" is used instead of "fentry" everywhere.
For example, the function kgr_get_fentry_loc() gets renamed to
kgr_get_function_address().

This patch does not modify the existing behavior on both supported
architectures: x86_64 and s390x. ftrace_function_to_fentry() always
returns the function address.

[*] Some instructions might be called before the ftrace handler
but they must be compatible with the epilogue of the new function.
An example is PPC where we first need to set TOC.

Signed-off-by: Petr Mladek <pmladek@suse.com>
Reviewed-by: Miroslav Benes <mbenes@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>