kgr: Safe way to avoid an infinite redirection
The slow stub currently always modifies the IP address. It sets it either
to the new or to the old function according to the current universe.

We call the _new_ function from the beginning. But we skip the size
of the ftrace call in the _old_ function to avoid an infinite loop.

It works on the currently supported architectures because the ftrace
handler call is the very first instruction in the function. But it
might cause problems on some other architectures. For example, TOC is
set before the ftrace handler call on PPC.

The size of the skipped area might be hard to compute. Fortunately,
we do not need to modify the IP address if we want to continue
with the original function (no redirection). Therefore we do not
need to know the position and size of the ftrace call at all.

This patch adds the check for the original function. It renames
"go_old" to "go_new" and inverts the logic. It makes it easier
to read the reworked final check.

Signed-off-by: Petr Mladek <pmladek@suse.com>
Reviewed-by: Miroslav Benes <mbenes@suse.cz>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>