kgr: kmemleak, really mark the kthread safe after an interrupt
kmemleak kthread could block kGraft patch finalization for few
minutes.  The problem is that it goes into the interruptible sleep
repeatedly until the entire original timeout is reached. We mark the
thread as safe outside this cycle.

This patch moves kgr_task_safe(t) into the wait cycle, so that it is
marked safe after the first interrupt.

Signed-off-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Jiri Slaby <jslaby@suse.cz>